backend/requirements.txt
# HireLiar Backend
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0

# ML
scikit-learn>=1.3.0
joblib>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
shap>=0.42.0

# Image
Pillow>=10.0.0

# URL & Domain Intelligence
requests>=2.31.0
python-whois>=0.9.0
beautifulsoup4>=4.12.0
dnspython>=2.4.0
aiohttp>=3.8.0
